+++ src/mock/point.js
@@ -28,7 +28,7 @@ const generateDescription = (city) => {
 const descriptions = CITIES.map((city) => generateDescription(city));
 
 export const generatePoint = () => {
-  const dateFrom = dayjs().add(getRandomNumber(-20, 20), 'day').format('YYYY-MM-DDTHH:mm:ss.SSS') + 'Z';
+  const dateFrom = dayjs().add(getRandomNumber(-28800, 28800), 'minute').format('YYYY-MM-DDTHH:mm:ss.SSS') + 'Z';
   const dateTo = dayjs(dateFrom).add(getRandomNumber(30, 160), 'minute').format('YYYY-MM-DDTHH:mm:ss.SSS') + 'Z';
   const type = getRandomElementFromArray(TYPES);
   const offersOfType = getArrayForType(TYPE_WITH_OFFERS, type.toLowerCase());
@@ -44,4 +44,4 @@ export const generatePoint = () => {
   };
 };
 
-export {descriptions };
+export { descriptions };
